+++ zingolib/src/testutils/chain_generics/with_assertions.rs
@@ -121,12 +121,10 @@ where
 
     timestamped_test_log("following proposal, preparing to unwind if an assertion fails.");
 
-    let server_height_at_send = BlockHeight::from(
-        crate::grpc_connector::get_latest_block(environment.lightserver_uri().unwrap())
-            .await
-            .unwrap()
-            .height as u32,
-    );
+    use zingo_netutils::Indexer as _;
+    let indexer = zingo_netutils::GrpcIndexer::new(environment.lightserver_uri().unwrap());
+    let server_height_at_send =
+        BlockHeight::from(indexer.get_latest_block().await.unwrap().height as u32);
     let last_known_chain_height = sender
         .wallet()
         .read()